Migrated CreateUserToDatabaseUsersTableAndRespondJsonOfUserHandler() to a new file sign_up_handler.
